+++ lbr_demos/lbr_demos_advanced_py/lbr_demos_advanced_py/lbr_base_position_command_node.py
@@ -34,7 +34,7 @@ class LBRBasePositionCommandNode(Node):
         )
         self._lbr_joint_position_command_pub = self.create_publisher(
             LBRJointPositionCommand,
-            "command/joint_position",
+            "command/lbr_joint_position_command",
             1,
         )