+++ lbr_fri_ros2/src/interfaces/torque_command.cpp
@@ -8,11 +8,17 @@ TorqueCommandInterface::TorqueCommandInterface(
 
 void TorqueCommandInterface::buffered_command_to_fri(fri_command_t_ref command,
                                                      const_idl_state_t_ref state) {
+  std::lock_guard<std::mutex> lock(command_mutex_);
   if (state.client_command_mode != KUKA::FRI::EClientCommandMode::TORQUE) {
-    std::string err = "Expected robot in '" +
-                      EnumMaps::client_command_mode_map(KUKA::FRI::EClientCommandMode::TORQUE) +
-                      "' command mode got '" +
-                      EnumMaps::client_command_mode_map(state.client_command_mode) + "'";
+    std::string err =
+        "Client side (configured via hardware.client_command_mode in lbr_system_config.yaml) "
+        "expected robot in '" +
+        EnumMaps::client_command_mode_map(KUKA::FRI::EClientCommandMode::TORQUE) +
+        "' command mode, but robot was in '" +
+        EnumMaps::client_command_mode_map(state.client_command_mode) +
+        "' command mode. Correct the configurations or run the robot in '" +
+        EnumMaps::client_command_mode_map(KUKA::FRI::EClientCommandMode::TORQUE) +
+        "' command mode.";
     RCLCPP_ERROR_STREAM(rclcpp::get_logger(LOGGER_NAME()),
                         ColorScheme::ERROR << err.c_str() << ColorScheme::ENDC);
     throw std::runtime_error(err);